requirements.txt
# --- Core App ---
streamlit==1.35.0

# --- Embeddings / NLP ---
sentence-transformers==2.6.1
transformers==4.41.2
torch==2.2.2
safetensors==0.4.3

# --- Vector Database ---
faiss-cpu==1.7.4

# --- Document Processing ---
PyPDF2==3.0.1
python-docx==1.1.0

# --- Optional Utilities ---
nltk==3.8.1
rouge-score==0.1.2

# --- Hugging Face Helpers ---
huggingface-hub==0.23.0
tokenizers==0.19.1

# --- General Utilities ---
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4